Debugging of the decoder network
Using a reference decoder my network was perfectly able to learn
attention. Therefore it seems that the bug is located in my decoder.

With this commit I will try to see if my decoder cell stack does work
with the reference attention mechanism that is known to work.

* In case it does work this would mean that the bug is located in my
attention cell and not inside my stacked decoder cells.
* In case it does not work this would mean that there is a flaw in my
stacked decoder cells. (Note that there would be still the possibility
of having a bug in my attention cell).